dreamer_sisaku/requirements.txt
absl-py==0.13.0
ale-py==0.7
asn1crypto==0.24.0
astunparse==1.6.3
atari-py==0.2.9
backcall==0.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cloudpickle==1.6.0
crafter==1.6.3
cryptography==2.1.4
cycler==0.11.0
dataclasses==0.8
decorator==5.1.0
dm-control==0.0.364896371
dm-env==1.5
dm-tree==0.1.6
flatbuffers==1.12
future==0.18.2
gast==0.3.3
glfw==2.2.0
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
gym==0.18.3
gym-minigrid==1.0.2
h5py==2.10.0
idna==2.6
imageio==2.9.0
importlib-metadata==4.5.0
importlib-resources==5.2.2
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
labmaze==1.0.5
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.3.4
numpy==1.19.4
oauthlib==3.1.1
opencv-python==4.5.5.64
opensimplex==0.3
opt-einsum==3.3.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.29
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
pyglet==1.5.15
Pygments==2.11.2
pygobject==3.26.1
PyOpenGL==3.1.5
pyparsing==2.4.7
python-apt==1.6.5+ubuntu0.6
python-dateutil==2.8.2
pyxdg==0.25
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scipy==1.5.4
SecretStorage==2.3.1
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.2
tensorflow-probability==0.12.2
termcolor==1.1.0
tqdm==4.62.3
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.5
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
